perf(simd): vectorized first-index recovery; fix RK3566 libuvc crash
- SIMD-firstIndexOf (SSE2 cmpeq+movemask / NEON vceqq+lane-OR) replaces
  the scalar index pass that dominated stats: x86 stats 7.0 -> 1.5
  us/frame (17x vs scalar). NEON stats kernel uses vpadalq widening sum.
- meson: skip -march=native when cross-compiling (host features would
  otherwise be baked into aarch64 code).
- subprojects: wrap patch for vendored libuvc - guard NULL isochronous
  packet buffers and NULL stream handle; RK3566 (dwc3) delivers NULL
  pkt buffers with actual_length > 0 which libuvc dereferenced (SIGTRAP).
- test_thermal: fix narrowing rejected by libc++.

Verified on ROCKNIX RK3566 (aarch64): NEON parity bit-exact on device,
stable 26 FPS under sway/wayland.

+++ src/tests/test_thermal.cpp
@@ -65,7 +65,7 @@ void TestSimdParity()
         // Colormap with several ranges, including extremes
         const uint16_t ranges[][2] = {
             {10000, 11000}, {0, 65535}, {40000, 40001}, {65000, 65535},
-            {data.front(), data.front() + 3}, {0, 256}, {65280, 65535}
+            {data.front(), static_cast<uint16_t>(data.front() + 3)}, {0, 256}, {65280, 65535}
         };
         for (auto& r : ranges)
         {
